# Data/Models/DTO/Account/ForgotPasswordRequest.cs
using System.ComponentModel.DataAnnotations;

namespace Models.DTO.Account
{
    public class ForgotPasswordRequest
    {
        [Required]
        [EmailAddress]
        public string Email { get; set; }
    }
}